Update README.md
Dear Colleagues, as the Imaging Facility Manager, I have been reflecting on the importance of establishing a robust theoretical framework to support imaging in microscopy. To address this, I have developed an approach that formalizes imaging through a set-theoretical perspective. I am pleased to share these insights in a concise paper titled: "Set-Theoretical Foundation of Imaging in Microscopy."